Factor out ProtoDecoderBase for extensibility
To support extensions like Jelly-Patch, it would be useful to factor out the common ProtoDecoder code to a trait that can be extended.

More info: https://github.com/Jelly-RDF/jelly-protobuf/issues/11

This is analogous to very similar changes in ProtoEncoder here: #278
